Updates to the cleaning script

+++ clean.sh
@@ -11,17 +11,21 @@
 #Set the working directory
 export CURRENTDIR=`pwd`
 
-#Remove BIN and OUT directories
+#Remove BIN and OUT directories, if they exist.
 cd $CURRENTDIR
-rm -r bin
-rm -r out
+if [ -d "bin" ]; then
+    rm -r bin
+fi
+if [ -d "out" ]; then
+    rm -r out
+fi
 
 #Switch to the kernel directory and clean up
 cd $CURRENTDIR/src/kernel
 make clean mrproper
 
 #Switch back to the working directory
-cd ../..
+cd $CURRENTDIR
 
 #EXIT
 exit
